pyproject.toml
[tool.poetry]
name = "rowbot"
version = "0.1.0"
description = "A bot serves the OneShot QQ group that's based on nonebot2 and uses cqhttp."
authors = ["KRedCell <krov_red_cell@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-alpha.13"
nonebot-adapter-cqhttp = "^2.0.0-alpha.13"
pydantic = "^1.8.2"
nonebot-plugin-apscheduler = "^0.1.2"
Pillow = "^8.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
